requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
idna==3.11
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.15.0
    # via
    #   altair
    #   plotly
numpy==2.4.2
    # via
    #   numpy-financial
    #   pandas
    #   pydeck
    #   streamlit
numpy-financial==1.0.0
    # via bc-real-estate (pyproject.toml)
packaging==26.0
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.1.0
    # via streamlit
plotly==6.5.2
    # via bc-real-estate (pyproject.toml)
protobuf==6.33.5
    # via streamlit
pyarrow==23.0.0
    # via streamlit
pydantic==2.12.5
    # via bc-real-estate (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.53.1
    # via bc-real-estate (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
